+++ src/Colors/ColorTranslator.cs
@@ -14,14 +14,14 @@ internal static class ColorTranslator
         FieldInfoList = typeof(SKColors).GetFields(BindingFlags.Static | BindingFlags.Public);
     }
 
-    internal static string HexFromName(string coloName)
+    internal static string HexFromName(string colorName)
     {
-        if (coloName.Equals("white", StringComparison.CurrentCultureIgnoreCase))
+        if (colorName.Equals("white", StringComparison.CurrentCultureIgnoreCase))
         {
             return "FFFFFF";
         }
 
-        var fieldInfo = FieldInfoList.First(fieldInfo => string.Equals(fieldInfo.Name, coloName, StringComparison.CurrentCultureIgnoreCase));
+        var fieldInfo = FieldInfoList.First(fieldInfo => string.Equals(fieldInfo.Name, colorName, StringComparison.CurrentCultureIgnoreCase));
         var color = (SKColor)fieldInfo.GetValue(null) !;
 
         return color.ToString();