+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog  
 
+## Version 0.65.1 - Unreleased
+🐞Fixed added picture width [#993](https://github.com/ShapeCrawler/ShapeCrawler/issues/993)  
+
+## Version 0.65.0 - 2025-04-11
+🍀Added `AsMarkdown()` to get presentation string in Markdown format [(example)](https://github.com/ShapeCrawler/ShapeCrawler/blob/af7042bcdf83e77ecf9b1457415af589cf7bda7a/ShapeCrawler.Examples/Presentation.cs#L7-L12)  
+
 ## Version 0.64.3 - 2025-03-25
 🐞Fixed updating table height [#552](https://github.com/ShapeCrawler/ShapeCrawler/issues/552)  
 🐞Fixed New Line issue [#960](https://github.com/ShapeCrawler/ShapeCrawler/issues/960)